+++ src/test/java/de/muspellheim/commons/util/EventBusTests.java
@@ -6,6 +6,7 @@
 package de.muspellheim.commons.util;
 
 import java.util.*;
+import java.util.concurrent.*;
 import java.util.function.*;
 
 import org.junit.jupiter.api.*;
@@ -19,23 +20,37 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 class EventBusTests {
 
-    private List<String> stringEvents = new ArrayList<>();
-    private List<Integer> integerEvents = new ArrayList<>();
-    private List<Number> numberEvents = new ArrayList<>();
-    private List<Number> doubleEvents = new ArrayList<>();
+    private List<String> stringEvents;
+    private List<Integer> integerEvents;
+    private List<Number> numberEvents;
+    private List<Number> doubleEvents;
+
+    private Phaser phaser;
     private Throwable exception;
 
+    @BeforeEach
+    void setUp() {
+        stringEvents = new ArrayList<>();
+        integerEvents = new ArrayList<>();
+        numberEvents = new ArrayList<>();
+        doubleEvents = new ArrayList<>();
+
+        phaser = new Phaser(1);
+    }
+
     @Test
-    void publish() {
+    void publish() throws Exception {
         // Given
         EventBus bus = EventBus.getDefault();
         bus.subscribe(String.class, this::consumeString);
         bus.subscribe(Integer.class, this::consumeInt);
 
         // When
+        phaser = new Phaser(3);
         bus.publish("Foo");
         bus.publish(42);
         bus.publish("Bar");
+        phaser.awaitAdvanceInterruptibly(0, 2, TimeUnit.SECONDS);
 
         // Then
         assertAll(
@@ -46,9 +61,9 @@ class EventBusTests {
     }
 
     @Test
-    void handlerThrowsException() {
+    void handlerThrowsException() throws Exception {
         // Given
-        Thread.currentThread().setUncaughtExceptionHandler((t, e) -> exception = e);
+        Thread.setDefaultUncaughtExceptionHandler((t, e) -> exception = e);
         EventBus bus = EventBus.getDefault();
         bus.subscribe(String.class, m -> {
             throw new RuntimeException("Foobar");
@@ -57,6 +72,7 @@ class EventBusTests {
 
         // When
         bus.publish("Foo");
+        TimeUnit.SECONDS.sleep(1);
 
         // Then
         assertAll(
@@ -66,7 +82,7 @@ class EventBusTests {
     }
 
     @Test
-    void unsubscribe() {
+    void unsubscribe() throws Exception {
         // Given
         EventBus bus = EventBus.getDefault();
         Consumer<String> subscriber = this::consumeString;
@@ -76,6 +92,7 @@ class EventBusTests {
         bus.publish("Foo");
         bus.unsubscribe(subscriber);
         bus.publish("Bar");
+        TimeUnit.SECONDS.sleep(1);
 
         // Then
         assertAll(
@@ -85,7 +102,7 @@ class EventBusTests {
     }
 
     @Test
-    void eventTypeHierarchy() {
+    void eventTypeHierarchy() throws Exception {
         // Given
         EventBus bus = EventBus.getDefault();
         Consumer<Integer> intSubscriber = this::consumeInt;
@@ -103,6 +120,7 @@ class EventBusTests {
         bus.publish(2.718);
         bus.unsubscribe(Integer.class, intSubscriber);
         bus.publish(7);
+        TimeUnit.SECONDS.sleep(1);
 
         // Then
         assertAll(
@@ -115,18 +133,22 @@ class EventBusTests {
 
     private void consumeString(String event) {
         stringEvents.add(event);
+        phaser.arrive();
     }
 
     private void consumeInt(int event) {
         integerEvents.add(event);
+        phaser.arrive();
     }
 
     private void consumeNumber(Number event) {
         numberEvents.add(event);
+        phaser.arrive();
     }
 
     private void consumeDouble(Number event) {
         doubleEvents.add(event);
+        phaser.arrive();
     }
 
 }