+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.6.0] - 2019-11-02
+
+### Changes
+
+*   Event bus delivers events asynchronously.
+
 ### Fixes
 
 *   Creating `About` when reading from class object.